BEX-1163 Throw warn about changes in serialization after migration to jackson-kotlin-module-2.15
jackson-kotlin-module contains a breaking change in the way how `is` functions are serialized in kotlin - they will now keep the `is` prefix. See https://github.com/FasterXML/jackson-module-kotlin/issues/670
In preparation to migrating to jackson-kotlin-module-2.15 we need to safeguard our existing serialized classes with @JsonProperty annotations explicitly stating the expected property name. To do that, we want to throw an exception if such annotation is absent on affected methods in non-production environments and log an error message in production environments - changes were added to BeanPropertyWriter to do that.